more embedded grid unit tests
