+++ tests/src/resources/config/log4j-embedded-grid.properties
@@ -7,7 +7,7 @@
 # you may not use this file except in compliance with the License.             #
 # You may obtain a copy of the License at                                      #
 #                                                                              #
-#    http://www.apache.org/licenses/LICENSE-2.0                                #
+# 	 http://www.apache.org/licenses/LICENSE-2.0                                #
 #                                                                              #
 # Unless required by applicable law or agreed to in writing, software          #
 # distributed under the License is distributed on an "AS IS" BASIS,            #
@@ -36,7 +36,7 @@ log4j.appender.JPPF.layout.ConversionPattern=%d [%-5p][%-20.20t][%c.%M(%L)]: %m%
 log4j.rootLogger=INFO, JPPF
 #log4j.rootLogger=DEBUG, JPPF
 
-log4j.logger.org.jppf.classloader.AbstractResourceProvider=DEBUG
+#log4j.logger.org.jppf.classloader.AbstractResourceProvider=DEBUG
 log4j.logger.org.jppf.client=DEBUG
 #log4j.logger.org.jppf.client.AbstractJPPFClient=DEBUG
 log4j.logger.org.jppf.client.AbstractJPPFClientConnection=TRACE
@@ -64,17 +64,25 @@ log4j.logger.org.jppf.execute=DEBUG
 #log4j.logger.org.jppf.management=DEBUG
 #log4j.logger.org.jppf.management.JPPFSystemInformation=TRACE
 #log4j.logger.org.jppf.management.JMXConnectionWrapper=INFO
+log4j.logger.org.jppf.management.spi.AbstractMBeanProviderManager=DEBUG
 #log4j.logger.org.jppf.classloader.ResourceProvider=DEBUG
-log4j.logger.org.jppf.load.balancer=DEBUG
-log4j.logger.org.jppf.serialization.JPPFSerialization$Factory=DEBUG
-log4j.logger.org.jppf.ssl=DEBUG
+#log4j.logger.org.jppf.load.balancer=DEBUG
+#log4j.logger.org.jppf.serialization.JPPFSerialization$Factory=DEBUG
+#log4j.logger.org.jppf.ssl=DEBUG
 #log4j.logger.org.jppf.utils.configuration=DEBUG
 #log4j.logger.org.jppf.utils.configuration.SubstitutionsHandler=TRACE
 log4j.logger.org.jppf.utils.JPPFDefaultUncaughtExceptionHandler=DEBUG
+log4j.logger.org.jppf.utils.concurrent.ConcurrentUtils=TRACE
 #log4j.logger.test.org.jppf.test.setup.common.TestInterceptor=DEBUG
 #log4j.logger.test.org.jppf.comm.interceptor.TestNetworkCommunicationInterceptor=DEBUG
 #log4j.logger.test.org.jppf.comm.interceptor.TestNetworkCommunicationInterceptor$SocketServer=DEBUG
 #log4j.logger.test.org.jppf.comm.interceptor.TestNetworkCommunicationInterceptor$ChannelServer=DEBUG
+log4j.logger.org.jppf.node=DEBUG
+log4j.logger.org.jppf.server.JPPFDriver=DEBUG
+log4j.logger.org.jppf.server.DriverInitializer=DEBUG
+log4j.logger.org.jppf.server.node=DEBUG
+log4j.logger.org.jppf.server.nio.client=DEBUG
+log4j.logger.org.jppf.server.nio.nodeserver=DEBUG
 #log4j.logger.SocketServer=DEBUG
 #log4j.logger.ChannelServer=DEBUG
 log4j.logger.org.jppf.test.setup.common.CountingJobListener=DEBUG
